moved requireToken middleware to separate file
